backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg[binary]
pydantic[email]
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
